Fix x64 client crashes on login scene load
- PetProcess::LoadData: copy pet speed data with sizeof(float) instead of
  sizeof(pointer), which corrupted the heap on x64.
- PetInfo::SetActions: reject invalid action counts to avoid allocating
  gigabytes of memory on malformed pet.bmd entries.
- CameraMove::LoadCameraWalkScript: read waypoint count as DWORD (4 bytes)
  because .cws files are x86 format; size_t on x64 reads 8 bytes and caused
  bad_alloc/abort when loading the login scene camera script.

+++ src/source/Camera/CameraMove.cpp
@@ -194,14 +194,14 @@ bool CCameraMove::LoadCameraWalkScript(const std::wstring& filename)
         return false;
     }
 
-    size_t waypointCount = 0;
-    if (fread(&waypointCount, sizeof(size_t), 1, fileHandle.get()) != 1)
+    DWORD waypointCount = 0;
+    if (fread(&waypointCount, sizeof(DWORD), 1, fileHandle.get()) != 1)
     {
         return false;
     }
 
     m_listWayPoint.reserve(waypointCount);
-    for (size_t index = 0; index < waypointCount; ++index)
+    for (DWORD index = 0; index < waypointCount; ++index)
     {
         auto waypoint = std::make_unique<WAYPOINT>();
         if (fread(waypoint.get(), sizeof(WAYPOINT), 1, fileHandle.get()) != 1)
